# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Restaurant Management System API",
        "description": "API documentation for Restaurant Management System Backend",
        "contact": {
            "email": "admin@restaurant.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Restaurant Management API Server"
        }
    ],
    "paths": {
...
        "/api/auth/dish-categories": {
            "get": {
                "tags": [
                    "DishCategories"
                ],
                "summary": "Get all dish categories",
                "description": "Retrieve all dish categories with pagination and optional search by name",
                "operationId": "3f9f840c91e79d88051d938be3e887a2",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 100
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Search by category name (partial match)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "đồ ăn hàn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dish categories retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Dish categories retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "string",
                                                                "example": "DC001"
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Món chính"
                                                            },
                                                            "desc": {
                                                                "type": "string",
                                                                "example": "Các món ăn chính"
                                                            },
                                                            "created_at": {
                                                                "type": "string",
                                                                "format": "date-time",
                                                                "example": "2025-10-03T12:00:00"
                                                            },
                                                            "updated_at": {
                                                                "type": "string",
                                                                "format": "date-time",
                                                                "example": "2025-10-03T12:00:00"
                                                            },
                                                            "dishes_count": {
                                                                "type": "integer",
                                                                "example": 8
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 50
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "DishCategories"
                ],
                "summary": "Create new dish category",
                "description": "Create a new dish category",
                "operationId": "975dd96b977688e8db3e3ac1336898c1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Khai vị"
                                    },
                                    "desc": {
                                        "type": "string",
                                        "example": "Các món khai vị"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Dish category created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/dish-categories/{id}": {
            "put": {
                "tags": [
                    "DishCategories"
                ],
                "summary": "Update dish category",
                "description": "Update an existing dish category",
                "operationId": "fd2a2eb8f6d2495a6e2eb47d1408f1d9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Dish Category ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Món chính"
                                    },
                                    "desc": {
                                        "type": "string",
                                        "example": "Các món chính"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Dish category updated successfully"
                    },
                    "404": {
                        "description": "Dish category not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "DishCategories"
                ],
                "summary": "Delete dish category",
                "description": "Delete a dish category. If any dishes are linked, deletion is prevented.",
                "operationId": "c2ff989b42e58a3aca0ac5b7ab61972f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Dish Category ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dish category deleted successfully"
                    },
                    "400": {
                        "description": "Cannot delete category with linked dishes"
                    },
                    "404": {
                        "description": "Dish category not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/dish-categories/get-name-list-dish-category": {
            "get": {
                "tags": [
                    "DishCategories"
                ],
                "summary": "Lấy danh sách tên danh mục món ăn",
                "description": "Trả về danh sách {id, name} của các danh mục món ăn",
                "operationId": "getNameListDishCategory",
                "responses": {
                    "200": {
                        "description": "Danh sách tên danh mục món ăn",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Món khai vị"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter JWT token in format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}